+++ x/profiles/types/keys.go
@@ -57,7 +57,7 @@ var (
 
 // DTagStoreKey turns a DTag into the key used to store the address associated with it into the store
 func DTagStoreKey(dTag string) []byte {
-	return append(DTagPrefix, []byte(dTag)...)
+	return append(DTagPrefix, []byte(strings.ToLower(dTag))...)
 }
 
 // DTagTransferRequestStoreKey turns an address to a key used to store a transfer request into the profiles store